refactor/#6 [refactor] : 중복된 Retrofit 코드 제거
